# server/nodejs/api/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0.4"
  title: "Syscoin API"
host: "localhost:8001"
...
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /certinfo:
    get:
      tags:
      - "Certificates"
      description: "Show stored values of a single certificate."
      operationId: "certinfo"
      parameters:
      - name: "guid"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Cert"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Certificates"
    x-swagger-router-controller: "rpc"
  /certfilter:
    get:
      tags:
      - "Certificates"
      description: "scan and filter certs"
      operationId: "certfilter"
      parameters:
      - name: "query"
        in: "query"
        description: "Generic filter query to pass into syscoinquery"
        required: true
        type: "string"
      - name: "count"
        in: "query"
        description: "The number of results to return"
        required: false
        type: "string"
      - name: "from"
        in: "query"
        description: "Show results from this GUID [from], empty means first"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CertIndex"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Certificates"
    x-swagger-router-controller: "rpc"
  /certhistory:
    get:
      tags:
      - "Certificates"
      description: "List all stored values of an cert."
      operationId: "certhistory"
      parameters:
      - name: "query"
        in: "query"
        description: "Generic filter query to pass into syscoinquery"
        required: true
        type: "string"
      - name: "count"
        in: "query"
        description: "The number of results to return"
        required: false
        type: "number"
      - name: "from"
        in: "query"
        description: "Show results from this GUID [from], empty means first"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CertHistoryIndex"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Certificates"
    x-swagger-router-controller: "rpc"
  /certnew:
    post:
      tags:
      - "Certificates"
      description: "Create a new Syscoin Certificate. Requires wallet passphrase to\
        \ be set with walletpassphrase call."
      operationId: "certnew"
      parameters:
      - in: "body"
        name: "request"
        required: true
        schema:
          $ref: "#/definitions/CertNewRequest"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Certificates"
    x-swagger-router-controller: "rpc"
  /certtransfer:
    post:
      tags:
      - "Certificates"
      description: "Transfer certificate from one user to another. Requires wallet\
        \ passphrase to be set with walletpassphrase call."
      operationId: "certtransfer"
      parameters:
      - in: "body"
        name: "request"
        required: true
        schema:
          $ref: "#/definitions/CertTransferRequest"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Certificates"
    x-swagger-router-controller: "rpc"
  /certupdate:
    post:
      tags:
      - "Certificates"
      description: "Perform an update on an certificate you control. Requires wallet\
        \ passphrase to be set with walletpassphrase call."
      operationId: "certupdate"
      parameters:
      - in: "body"
        name: "request"
        required: true
        schema:
          $ref: "#/definitions/CertUpdateRequest"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "Certificates"
    x-swagger-router-controller: "rpc"
...
securityDefinitions:
  token:
    type: "apiKey"
    name: "token"
    in: "header"
definitions:
...
  ErrorResponse:
    required:
    - "message"
    properties:
      message:
        type: "string"
...
  Cert:
    properties:
      _id:
        type: "string"
      txid:
        type: "string"
      height:
        type: "number"
      title:
        type: "string"
      time:
        type: "number"
      publicvalue:
        type: "string"
      category:
        type: "string"
      alias:
        type: "string"
      access_flags:
        type: "number"
      expires_on:
        type: "number"
      expired:
        type: "boolean"
    example:
      publicvalue: "aeiou"
      expired: true
      expires_on: 5.962133916683182377482808078639209270477294921875
      txid: "aeiou"
      alias: "aeiou"
      _id: "aeiou"
      time: 6.02745618307040320615897144307382404804229736328125
      access_flags: 1.46581298050294517310021547018550336360931396484375
      title: "aeiou"
      category: "aeiou"
      height: 0.80082819046101150206595775671303272247314453125
  CertIndex:
    properties:
      _id:
        type: "string"
      title:
        type: "string"
      height:
        type: "number"
      category:
        type: "string"
      alias:
        type: "string"
    example:
      alias: "aeiou"
      _id: "aeiou"
      title: "aeiou"
      category: "aeiou"
      height: 0.80082819046101150206595775671303272247314453125
  CertHistoryIndex:
    properties:
      _id:
        type: "string"
      cert:
        type: "string"
      height:
        type: "number"
      time:
        type: "number"
      title:
        type: "string"
      publicvalue:
        type: "string"
      category:
        type: "string"
      alias:
        type: "string"
      access_flags:
        type: "number"
      op:
        type: "string"
    example:
      op: "aeiou"
      publicvalue: "aeiou"
      alias: "aeiou"
      cert: "aeiou"
      _id: "aeiou"
      time: 6.02745618307040320615897144307382404804229736328125
      access_flags: 1.46581298050294517310021547018550336360931396484375
      title: "aeiou"
      category: "aeiou"
      height: 0.80082819046101150206595775671303272247314453125
...
  CertNewRequest:
    required:
    - "alias"
    - "public"
    - "title"
    properties:
      alias:
        type: "string"
        description: "An alias you own."
      title:
        type: "string"
        description: "title, 256 characters max."
      public:
        type: "string"
        description: "public data, 512 characters max."
      category:
        type: "string"
        description: "category, 256 characters max. Defaults to certificates;"
      witness:
        type: "string"
        description: "Witness alias name that will sign for web-of-trust notarization\
          \ of this transaction."
  CertTransferRequest:
    required:
    - "alias"
    - "certkey"
    properties:
      certkey:
        type: "string"
        description: "Certificate guidkey."
      alias:
        type: "string"
        description: "Alias to transfer this certificate to."
      public:
        type: "string"
        description: "Public certificate data, 512 characters max."
      accessflags:
        type: "string"
        description: "Set new access flags for new owner for this certificate, 0 for\
          \ read-only, 1 for edit, 2 for edit and transfer access."
      witness:
        type: "string"
        description: "Witness alias name that will sign for web-of-trust notarization\
          \ of this transaction."
  CertUpdateRequest:
    required:
    - "guid"
    properties:
      guid:
        type: "string"
        description: "Certificate guidkey."
      title:
        type: "string"
        description: "Certificate title, 256 characters max."
      public:
        type: "string"
        description: "Public certificate data, 512 characters max."
      category:
        type: "string"
        description: "Category, 256 characters max. Defaults to certificates."
      witness:
        type: "string"
        description: "Witness alias name that will sign for web-of-trust notarization\
          \ of this transaction."
...